pyproject.toml
[tool.poetry]
name = "flirtools"
version = "0.1.0"
description = "Tools for using FLIR cameras"
authors = ["spencer <spencer.hallyburton@duke.edu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
numpy = ">=1.20, <1.26"
opencv-python = ">4.7, <4.9.0.80"
Pillow = "^9"
numpy-quaternion = "^2022.4"
spinnaker_python = [
    { url = "https://g-b0ef78.1d0d8d.03c0.data.globus.org/packages/pyspin/spinnaker_python-3.1.0.79-cp310-cp310-linux_x86_64.whl", markers = "platform_machine == 'x86_64'" },
    { url = "https://g-b0ef78.1d0d8d.03c0.data.globus.org/packages/pyspin/spinnaker_python-3.1.0.79-cp310-cp310-linux_aarch64.whl", markers = "platform_machine == 'aarch64'" }
]

[tool.poetry.group.test.dependencies]
pytest = "^7"
pytest-cov = "^4"
bandit = "^1.7"
black = "~22.12.0"
mypy = "^0"
isort = "^5.11"
flake8 = "^5.0"
autoflake = "^1"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q -s"
testpaths = [
    "tests",
    "integration",
]

[build-system]
requires = [
    "poetry-core",
]
build-backend = "poetry.core.masonry.api"
